Move storage and transformations out of common
