+++ src/data/transformations/__tests__/EventsTransformationD2Api.spec.ts
@@ -1,6 +1,6 @@
 import _ from "lodash";
-import { Transformation } from "../../../domain/common/entities/Transformation";
 import { EventsPackage } from "../../../domain/events/entities/EventsPackage";
+import { Transformation } from "../../../domain/transformations/entities/Transformation";
 import { TransformationD2ApiRepository } from "../TransformationD2ApiRepository";
 import { D2EventsPackage } from "../types";
