+++ src/domain/aggregated/usecases/AggregatedSyncUseCase.ts
@@ -11,7 +11,6 @@ import {
     mapCategoryOptionCombo,
     mapOptionValue,
 } from "../../../utils/synchronization";
-import { TransformationRepository } from "../../common/repositories/TransformationRepository";
 import { Instance as InstanceEntity } from "../../instance/entities/Instance";
 import { InstanceRepository } from "../../instance/repositories/InstanceRepository";
 import { GenericSyncUseCase } from "../../synchronization/usecases/GenericSyncUseCase";
@@ -20,6 +19,7 @@ import {
     cleanObjectDefault,
     cleanOrgUnitPath,
 } from "../../synchronization/utils";
+import { TransformationRepository } from "../../transformations/repositories/TransformationRepository";
 import { AggregatedPackage } from "../entities/AggregatedPackage";
 import { DataValue } from "../entities/DataValue";
 import { AggregatedRepository } from "../repositories/AggregatedRepository";
