+++ src/domain/events/usecases/EventsSyncUseCase.ts
@@ -15,7 +15,6 @@ import {
 import { DataValue } from "../../aggregated/entities/DataValue";
 import { AggregatedRepository } from "../../aggregated/repositories/AggregatedRepository";
 import { AggregatedSyncUseCase } from "../../aggregated/usecases/AggregatedSyncUseCase";
-import { TransformationRepository } from "../../common/repositories/TransformationRepository";
 import { Instance as InstanceEntity } from "../../instance/entities/Instance";
 import { InstanceRepository } from "../../instance/repositories/InstanceRepository";
 import {
@@ -23,6 +22,7 @@ import {
     SyncronizationPayload,
 } from "../../synchronization/usecases/GenericSyncUseCase";
 import { buildMetadataDictionary, cleanOrgUnitPath } from "../../synchronization/utils";
+import { TransformationRepository } from "../../transformations/repositories/TransformationRepository";
 import { EventsPackage } from "../entities/EventsPackage";
 import { ProgramEvent } from "../entities/ProgramEvent";
 import { ProgramEventDataValue } from "../entities/ProgramEventDataValue";
